Cross build for play 2.5 and 2.6, and scala 2.12
