+++ .travis.yml
@@ -6,12 +6,17 @@ cache:
     - $HOME/.ivy2/cache
     - $HOME/.sbt/boot/
 
-scala:
-- 2.11.7
-
 jdk:
-- oraclejdk8
+- openjdk8
 
+matrix:
+  include:
+    - scala: 2.11.12
+      env: PLAY_VERSION=2.5
+    - scala: 2.11.12
+      env: PLAY_VERSION=2.6
+    - scala: 2.12.8
+      env: PLAY_VERSION=2.6
 
 script:
   - sbt ++$TRAVIS_SCALA_VERSION -J-XX:ReservedCodeCacheSize=256M test
