+++ project/PlayCrossCompilation.scala
@@ -1,0 +1,4 @@
+import uk.gov.hmrc.playcrosscompilation.AbstractPlayCrossCompilation
+import uk.gov.hmrc.playcrosscompilation.PlayVersion.Play25
+
+object PlayCrossCompilation extends AbstractPlayCrossCompilation(defaultPlayVersion = Play25)
